// src/lib/enum.ts
// ROLES
export enum Role {
    ADMIN = 'ADMIN',
    SUPER_ADMIN = 'SUPER_ADMIN',
    DOCTOR = 'DOCTOR',
    PATIENT = 'PATIENT',
    STAFF = 'STAFF'
}

export enum Gender {
    MALE = 'MALE',
    FEMALE = 'FEMALE',
    OTHER = 'OTHER'
}

export enum MaritalStatus {
    CELIBATAIRE = 'CELIBATAIRE',
    MARIE = 'MARIE(E)',
    DIVORCE = 'DIVORCE(E)',
    VEUF = 'VEUF(VE)'
}